# secp256k1: Add fixed-precision group order type.