… dcrec/secp256k1/modnscalar.go
… dcrec/secp256k1/modnscalar_test.go