**This requires #2059**.

This introduces a specialized type for handling arithmetic modulo the group order which is significantly faster than using big integers from the standard library

The implementation is also constant time, whereas big integers from the standard library are not, which helps protect against side channel attacks in certain scenarios.

Comprehensive tests with 100% coverage and a full suite of benchmarks are included.

Due to the fact the code relies on a lot of low-level bit manipulations to achieve the speed and constant time characteristics, it is not the easiest code to review for correctness, so significant effort has been put into attempting to clearly explain the details via comments to help with review.

For now, this merely introduces the type without modifying any of the code to make use of it, and it also does not yet include the ability to perform inversions which will be required in order to switch over to it.  That capability will be added in a future commit.

The following benchmarks show a comparison between the new specialized type and generic big ints for various operations:

```
benchmark                  old ns/op   new ns/op     delta
------------------------------------------------------------
BenchmarkModN              267         18.9          -92.92%
BenchmarkZero              7.28        0.64          -91.25%
BenchmarkIsZero            0.32        0.32          +0.00%
BenchmarkEquals            9.14        5.58          -38.95%
BenchmarkAddModN           305         23.8          -92.20%
BenchmarkMulModN           457         237           -48.14%
BenchmarkSquareModN        459         238           -48.15%
BenchmarkNegateModN        295         9.82          -96.67%
BenchmarkIsOverHalfOrder   9.25        8.55          -7.57%

benchmark                  old allocs   new allocs   delta
------------------------------------------------------------
BenchmarkModN              2            0            -100.00%
BenchmarkZero              0            0            +0.00%
BenchmarkIsZero            0            0            +0.00%
BenchmarkEquals            0            0            +0.00%
BenchmarkAddModN           2            0            -100.00%
BenchmarkMulModN           2            0            -100.00%
BenchmarkSquareModN        2            0            -100.00%
BenchmarkNegateModN        2            0            -100.00%
BenchmarkIsOverHalfOrder   0            0            +0.00%
```