@davecgh 2020-02-08
Test coverage:
```
$ go test -coverprofile=cov.out && go tool cover -func=cov.out | grep modnscalar | sed -e "s#github.com/decred/dcrd/dcrec/secp256k1/v3/##"
modnscalar.go:142:              String                  100.0%
modnscalar.go:152:              Set                     100.0%
modnscalar.go:160:              Zero                    100.0%
modnscalar.go:172:              IsZero                  100.0%
modnscalar.go:184:              SetInt                  100.0%
modnscalar.go:191:              constantTimeEq          100.0%
modnscalar.go:196:              constantTimeNotEq       100.0%
modnscalar.go:201:              constantTimeLess        100.0%
modnscalar.go:206:              constantTimeLessOrEq    100.0%
modnscalar.go:211:              constantTimeGreater     100.0%
modnscalar.go:216:              constantTimeGreaterOrEq 100.0%
modnscalar.go:221:              constantTimeMin         100.0%
modnscalar.go:227:              overflows               100.0%
modnscalar.go:258:              reduce256               100.0%
modnscalar.go:297:              SetBytes                100.0%
modnscalar.go:328:              SetByteSlice            100.0%
modnscalar.go:340:              PutBytes                100.0%
modnscalar.go:382:              Bytes                   100.0%
modnscalar.go:389:              IsOdd                   100.0%
modnscalar.go:395:              Equals                  100.0%
modnscalar.go:410:              Add2                    100.0%
modnscalar.go:439:              Add                     100.0%
modnscalar.go:450:              Add                     100.0%
modnscalar.go:460:              Rsh32                   100.0%
modnscalar.go:468:              reduce385               100.0%
modnscalar.go:597:              reduce512               100.0%
modnscalar.go:761:              Mul2                    100.0%
modnscalar.go:922:              Mul                     100.0%
modnscalar.go:931:              SquareVal               100.0%
modnscalar.go:948:              Square                  100.0%
modnscalar.go:957:              NegateVal               100.0%
modnscalar.go:997:              Negate                  100.0%
modnscalar.go:1075:             IsOverHalfOrder         100.0%
```